Moved rerun to own file, needs to be fixed in Makefile and tested.
Moved a number of ifdefs to module.595 and init_par.f95. Now, variables for sediment and streamfunctions are initialized in each module. Needs to be tested.

New option to save tt instead of ints. Binwrite is change so that time always is saved as 8byte bvariable. This is controlled by twritetype.

Improved jd as a global time variable.

Moved start1d, start2d, start3d etc to init_par.f95. These variables shouldn't be set in readfile.f95 or setupgrid.f95.

Moved the init aof a number of varaibles to init_par.f95 to improve readibility of loop.f95

Added non-allocated mask variable to mod_grid.

Added the strick flag to the error check in loop.f95. If set, the code will stop for non-chritical errors such as particle on land.

Added a test in loop.f95 to make sure that dxyz always is positive.



git-svn-id: http://svn.tracmass.org/trunk@255 2ee5d717-d111-0410-bb1d-d2df4af4126a

git-svn-id: https://github.com/TRACMASS/tracmass.git@216 dd566b57-2d5a-cdae-10d7-ce57b7db7bff